+++ core/common/permissions/src/main/AndroidManifest.xml
