+++ core/common/permissions/src/main/java/com/buzbuz/smartautoclicker/core/common/permissions/model/PermissionPostNotification.kt
@@ -14,20 +14,23 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-plugins {
-    alias(libs.plugins.buzbuz.androidLibrary)
-    alias(libs.plugins.buzbuz.hilt)
-}
+package com.buzbuz.smartautoclicker.core.common.permissions.model
 
-android {
-    namespace = "com.buzbuz.smartautoclicker.feature.permissions"
-    buildFeatures.viewBinding = true
-}
+import android.Manifest
+import android.annotation.SuppressLint
+import android.app.NotificationManager
+import android.content.Context
+import android.os.Build
 
-dependencies {
-    implementation(libs.androidx.appCompat)
-    implementation(libs.androidx.fragment.ktx)
-    implementation(libs.google.material)
+@SuppressLint("InlinedApi")
+data object PermissionPostNotification : Permission.Dangerous(), Permission.Optional, Permission.ForApiRange {
 
-    implementation(project(":core:common:ui"))
+    override val fromApiLvl: Int
+        get() = Build.VERSION_CODES.TIRAMISU
+
+    override val permissionString: String
+        get() = Manifest.permission.POST_NOTIFICATIONS
+
+    override fun isGranted(context: Context): Boolean =
+        context.getSystemService(NotificationManager::class.java).areNotificationsEnabled()
 }
