+++ core/common/permissions/src/main/res/layout/dialog_permission.xml
