+++ core/common/permissions/src/main/res/values-fr/strings.xml
