+++ core/common/permissions/src/main/res/values-pt-rBR/strings.xml
