+++ core/common/permissions/src/main/res/values-zh/strings.xml
