+++ core/common/permissions/src/main/res/values/strings.xml
