+++ core/common/permissions/src/main/res/values/styles.xml
