+++ settings.gradle.kts
@@ -30,6 +30,7 @@ include(":core:common:base")
 include(":core:common:bitmaps")
 include(":core:common:display")
 include(":core:common:overlays")
+include(":core:common:permissions")
 include(":core:common:quality")
 include(":core:common:ui")
 include(":core:dumb")
@@ -39,7 +40,6 @@ include(":core:smart:domain")
 include(":core:smart:processing")
 
 include(":feature:backup")
-include(":feature:permissions")
 include(":feature:notifications")
 include(":feature:quick-settings-tile")
 include(":feature:revenue")
