+++ smartautoclicker/build.gradle.kts
@@ -98,7 +98,7 @@ dependencies {
     implementation(project(":core:smart:processing"))
     implementation(project(":feature:backup"))
     implementation(project(":feature:notifications"))
-    implementation(project(":feature:permissions"))
+    implementation(project(":core:common:permissions"))
     implementation(project(":feature:quick-settings-tile"))
     implementation(project(":feature:revenue"))
     implementation(project(":feature:smart-config"))
